+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Additional Calypso requirement related to the transaction audit data (issue [#68]).
 ### Upgraded
 - "Calypsonet Terminal Calypso API" to version `1.2.0` (issue [#66]).
+### Fixed
+- Null Pointer Exception raised when the "open secure session" command returns an unexpected status word [#70]
 
 ## [2.1.0] - 2022-02-01
 ### Added
@@ -66,6 +68,7 @@ It also brings many major API changes.
 [2.0.1]: https://github.com/eclipse/keyple-card-calypso-java-lib/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/eclipse/keyple-card-calypso-java-lib/releases/tag/2.0.0
 
+[#70]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/70
 [#68]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/68
 [#66]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/66
 [#64]: https://github.com/eclipse/keyple-card-calypso-java-lib/issues/64